Concept change #3
ServerID is used to know if a card is "new", so if a card gets moved to
another addressbook, the serverID needs to be stripped. However during
sync, the server may push new cards with serverID which of course we
need to keep. To seperate these events, the sync code adds the new card
without serverID and re-adds the serverID by modifying the added card.
This way, the onItemAdd listener can strip off any serverID found
(because they are not added during sync but by user action) and
everything should be fine.