propper handling of server side deletes
If a folder is synced and deleted on the server, the local copy gets an
"offline" suffix, but it is not deleted at the next sync.

Outlook/O365 does not directly delete but moves folders into trash (but
does not provide an undelete...). These trashed folders can now be
hidden in the account manager. Switch is still hardcoded.

Todo:
- add switch to UI
- allow undelete