fix folder delete feature
This is what I actually had in mind, when adding this: Some EAS service
provider like O365 do not purge a folder, if deleted in the web
interface, but move it to trash. That trash folder is not visible in the
web interface. However  TbSync can and will show such deleted folders.
The user can purge those folders to get rid of them or export them, if
it was deleted by mistake. It would also be possible to implement an
"undelete" option...