chore: update strict max version to `139.*`
Thunderbird `139.0` was released a month ago. This extension specifies
a strict max version of `138.*`, which prevents users on `139.*` from
installing it by making it inaccessible on the Thunderbird Add-ons site.

The update does not seem to interfere with the function of this extension,
so this commit simply updates the max version to `139.*` to ensure users
can install it on the latest Thunderbird version.